@@ -1,9 +1,9 @@
 module go.breu.io/quantm
 
-go 1.22
+go 1.23
 
 require (
-	cloud.google.com/go/logging v1.10.0
+	cloud.google.com/go/logging v1.11.0
 	github.com/GoogleCloudPlatform/opentelemetry-operations-go/exporter/trace v1.24.0
 	github.com/GoogleCloudPlatform/opentelemetry-operations-go/propagator v0.48.0
 	github.com/Guilospanck/gocqlxmock v1.0.1
